+++ BestNBAApp/BestNBAApp/app/src/main/AndroidManifest.xml
@@ -18,7 +18,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".SearchResults"></activity>
+        <activity android:name=".SearchResults" />
+        <activity android:name=".Main2Activity"></activity>
     </application>
 
 </manifest>
