+++ BestNBAApp/BestNBAApp/app/src/main/java/com/mai/airwi/bestnbaapp/MainActivity.java
@@ -1,68 +1,42 @@
 package com.mai.airwi.bestnbaapp;
 
 //import android.app.Fragment;
-import android.support.v4.app.Fragment;
-import android.support.annotation.NonNull;
-import android.support.design.internal.BottomNavigationItemView;
-import android.support.design.widget.BottomNavigationView;
+import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
-import android.widget.CompoundButton;
 import android.widget.EditText;
-import android.widget.Switch;
-import android.widget.TextView;
-
-import com.android.volley.Request;
-import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.StringRequest;
-import com.android.volley.toolbox.Volley;
-
-import java.io.IOException;
-import java.net.URLEncoder;
-import java.util.*;
 
 public class MainActivity extends AppCompatActivity {
+
+    EditText username;
+    EditText password;
+    Button logInButton;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        BottomNavigationView bottomNav = findViewById(R.id.bottom_navigation);
-        bottomNav.setOnNavigationItemSelectedListener(navListener);
+        username = (EditText)findViewById(R.id.usernameEditText);
+        password = (EditText)findViewById(R.id.passwordEditText);
+        logInButton = (Button)findViewById(R.id.loginButton);
 
-        getSupportFragmentManager().beginTransaction().replace(R.id.fragment_container,
-                new SearchFragment()).commit();
+        logInButton.setOnClickListener(new View.OnClickListener(){
+            @Override
+            public void onClick(View view) {
+                validateAndOpen(username.getText().toString(), password.getText().toString());
+            }
+        });
     }
 
-    private BottomNavigationView.OnNavigationItemSelectedListener navListener =
-            new BottomNavigationView.OnNavigationItemSelectedListener() {
-                @Override
-                public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-                    Fragment selectedFragment = null;
+    public void validateAndOpen(String userName, String password){
 
-                    switch(item.getItemId()){
-                        case R.id.nav_search:
-                            selectedFragment = new SearchFragment();
-                            break;
-                        case R.id.nav_basket:
-                            selectedFragment = new BasketFragment();
-                            break;
-                        case R.id.nav_analyze:
-                            selectedFragment = new AnalyzeFragment();
-                            break;
-                    }
+        if(((userName.equals("admin")) && (password.equals("12345")))){
+            Intent intent = new Intent(this,Main2Activity.class);
+            startActivity(intent);
+        }
+    }
 
-                    getSupportFragmentManager().beginTransaction().replace(R.id.fragment_container,
-                            selectedFragment).commit();
-
-                    return true;
-                }
-            };
 }
